Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -10,8 +10,8 @@
 // }
 #include <stdio.h>
 int main(){
-    float f;
-    scanf("%f",&f);
-    printf("You entered: %.4f",f);
+    double f;
+    scanf("%d",&f);
+    printf("You entered: %d",f);
     return 0;
 }
